move non-error tests to correct location
