+++ src/parser/tests.rs
@@ -1,10 +1,17 @@
 use crate::ast::{SourceRange, Statement};
 
 // Copyright (c) 2020 Ghaith Hachem and Mathias Rieder
+mod container_parser_tests;
 mod control_parser_tests;
 mod expressions_parser_tests;
+mod function_parser_tests;
+mod initializer_parser_tests;
 mod misc_parser_tests;
 mod parse_errors;
+mod program_parser_tests;
+mod statement_parser_tests;
+mod type_parser_tests;
+mod variable_parser_tests;
 
 pub fn lex(source: &str) -> crate::lexer::ParseSession {
     crate::lexer::lex(source)
